READY BEFORE PLAY STORE

+++ app/src/main/java/com/clerodri/binnacle/home/domain/model/Home.kt
@@ -4,6 +4,6 @@ data class Home(
     val currentIndex: Int,
     val isStarted: Boolean,
     val elapsedSeconds: Long,
-    val roundId: Int,
+    val roundId: Long,
     val checkId: Int
 )
